# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mystics Audit — Cloud-Based Accounting & GST Compliance SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List chart of accounts
      parameters:
        - { name: type, in: query, schema: { type: string } }
        - { name: group, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a ledger account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account with ledger entries
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Account with ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountLedger"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/trial-balance:
    get:
      operationId: getTrialBalance
      tags: [accounts]
      summary: Get trial balance for a date
      parameters:
        - { name: date, in: query, schema: { type: string } }
      responses:
        "200":
          description: Trial balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialBalance"

  # JOURNAL ENTRIES
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List vendor bills
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: vendorId, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorBill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create vendor bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorBillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorBill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get vendor bill
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Vendor bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorBill"
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update vendor bill
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorBillUpdate"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorBill"
  /bills/{id}/pay:
    post:
      operationId: payBill
      tags: [bills]
      summary: Record payment for a vendor bill
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillPaymentInput"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorBill"

  # BANK ACCOUNTS
  /bank/accounts:
    get:
      operationId: listBankAccounts
      tags: [bank]
      summary: List bank accounts with balances
      responses:
        "200":
          description: Bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank]
      summary: Add bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank/accounts/{id}/transactions:
    get:
      operationId: getBankTransactions
      tags: [bank]
      summary: Get bank transactions for reconciliation
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Bank transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankTransaction"
  /bank/accounts/{id}/reconcile:
    post:
      operationId: reconcileBankTransaction
      tags: [bank]
      summary: Match bank transaction to journal entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileInput"
      responses:
        "200":
          description: Reconciled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"

  # EXPENSES
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        type: { type: string }
        group: { type: string }
        subGroup: { type: ["string", "null"] }
        normalBalance: { type: string }
        isParty: { type: boolean }
        isBank: { type: boolean }
        isCash: { type: boolean }
        isActive: { type: boolean }
        openingBalance: { type: number }
        currentBalance: { type: number }
        description: { type: ["string", "null"] }
      required: [id, code, name, type, group, normalBalance, isParty, isBank, isCash, isActive, openingBalance, currentBalance]
    AccountInput:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        type: { type: string }
        group: { type: string }
        subGroup: { type: string }
        normalBalance: { type: string }
        isParty: { type: boolean }
        isBank: { type: boolean }
        isCash: { type: boolean }
        openingBalance: { type: number }
        description: { type: string }
      required: [code, name, type, group, normalBalance]
    AccountUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    AccountLedger:
      type: object
      properties:
        account: { $ref: "#/components/schemas/Account" }
        openingBalance: { type: number }
        closingBalance: { type: number }
        totalDebit: { type: number }
        totalCredit: { type: number }
        entries: { type: array, items: { $ref: "#/components/schemas/LedgerEntry" } }
      required: [account, openingBalance, closingBalance, totalDebit, totalCredit, entries]
    LedgerEntry:
      type: object
      properties:
        id: { type: integer }
        date: { type: string }
        voucherNo: { type: string }
        narration: { type: ["string", "null"] }
        debit: { type: number }
        credit: { type: number }
        balance: { type: number }
        type: { type: string }
      required: [id, date, voucherNo, debit, credit, balance, type]
    TrialBalance:
      type: object
      properties:
        asOf: { type: string }
        totalDebit: { type: number }
        totalCredit: { type: number }
        isBalanced: { type: boolean }
        lines: { type: array, items: { $ref: "#/components/schemas/TrialBalanceLine" } }
      required: [asOf, totalDebit, totalCredit, isBalanced, lines]
    TrialBalanceLine:
      type: object
      properties:
        accountCode: { type: string }
        accountName: { type: string }
        group: { type: string }
        debit: { type: number }
        credit: { type: number }
      required: [accountCode, accountName, group, debit, credit]

    # JOURNAL ENTRY
...
    VendorBill:
      type: object
      properties:
        id: { type: integer }
        billNo: { type: string }
        vendorInvoiceNo: { type: ["string", "null"] }
        vendorId: { type: integer }
        vendorName: { type: string }
        date: { type: string }
        dueDate: { type: string }
        poId: { type: ["integer", "null"] }
        grnId: { type: ["integer", "null"] }
        status: { type: string }
        taxableAmount: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        tdsAmount: { type: number }
        totalAmount: { type: number }
        paidAmount: { type: number }
        balanceDue: { type: number }
        isMsmeVendor: { type: boolean }
        msmeBreachDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/BillLine" } }
      required: [id, billNo, vendorId, vendorName, date, dueDate, status, taxableAmount, cgst, sgst, igst, tdsAmount, totalAmount, paidAmount, balanceDue, isMsmeVendor, createdAt, lines]
    BillLine:
      type: object
      properties:
        id: { type: integer }
        description: { type: string }
        hsnSac: { type: string }
        quantity: { type: number }
        unit: { type: string }
        rate: { type: number }
        gstRate: { type: number }
        taxableValue: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        lineTotal: { type: number }
      required: [id, description, hsnSac, quantity, unit, rate, gstRate, taxableValue, cgst, sgst, igst, lineTotal]
    VendorBillInput:
      type: object
      properties:
        vendorId: { type: integer }
        vendorInvoiceNo: { type: string }
        date: { type: string }
        dueDate: { type: string }
        poId: { type: integer }
        grnId: { type: integer }
        notes: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/BillLineInput" } }
      required: [vendorId, date, dueDate, lines]
    BillLineInput:
      type: object
      properties:
        description: { type: string }
        hsnSac: { type: string }
        quantity: { type: number }
        unit: { type: string }
        rate: { type: number }
        gstRate: { type: number }
      required: [description, hsnSac, quantity, unit, rate, gstRate]
    VendorBillUpdate:
      type: object
      properties:
        vendorInvoiceNo: { type: string }
        dueDate: { type: string }
        notes: { type: string }
    BillPaymentInput:
      type: object
      properties:
        date: { type: string }
        amount: { type: number }
        paymentMode: { type: string }
        bankAccountId: { type: integer }
        referenceNo: { type: string }
        narration: { type: string }
      required: [date, amount, paymentMode, bankAccountId]

    # BANK
    BankAccount:
      type: object
      properties:
        id: { type: integer }
        accountName: { type: string }
        bankName: { type: string }
        accountNo: { type: string }
        ifsc: { type: string }
        accountType: { type: string }
        branch: { type: ["string", "null"] }
        balance: { type: number }
        isActive: { type: boolean }
        lastReconciled: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, accountName, bankName, accountNo, ifsc, accountType, balance, isActive, createdAt]
    BankAccountInput:
      type: object
      properties:
        accountName: { type: string }
        bankName: { type: string }
        accountNo: { type: string }
        ifsc: { type: string }
        accountType: { type: string }
        branch: { type: string }
        openingBalance: { type: number }
      required: [accountName, bankName, accountNo, ifsc, accountType]
    BankTransaction:
      type: object
      properties:
        id: { type: integer }
        bankAccountId: { type: integer }
        date: { type: string }
        description: { type: string }
        debit: { type: number }
        credit: { type: number }
        balance: { type: number }
        referenceNo: { type: ["string", "null"] }
        status: { type: string }
        matchedJournalId: { type: ["integer", "null"] }
        matchConfidence: { type: ["number", "null"] }
      required: [id, bankAccountId, date, description, debit, credit, balance, status]
    ReconcileInput:
      type: object
      properties:
        bankTransactionId: { type: integer }
        journalId: { type: integer }
      required: [bankTransactionId, journalId]

    # EXPENSES
...